pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metagen"
version = "0.2.0"
description = "MetaGen: a universal spec-to-model synthesizer"
readme = "README.md"
authors = [{ name = "MetaGen Research", email = "research@metagen.local" }]
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "jinja2>=3.1.3",
  "matplotlib>=3.8.0",
  "rich>=13.7.0",
  "click>=8.1.7",
]

[project.optional-dependencies]
torch = ["torch>=2.2.0"]
dev = ["ruff>=0.3.0", "pytest>=8.0.0", "pytest-cov>=4.1.0", "torch>=2.2.0"]

[project.scripts]
metagen = "metagen.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["metagen"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_network: marks tests requiring network access",
    "requires_gpu: marks tests requiring CUDA GPU",
    "integration: marks integration tests",
    "unit: marks unit tests",
]
